requirements.txt
django
djangorestframework
PyYAML 
uritemplate
Markdown
Pygments
django-filter
django-guardian
google-generativeai
django-cors-headers==4.3.1
asgiref==3.8.1
sqlparse==0.5.0
tzdata==2024.1